refactor: de-dup some test setup code
